Added the Stradavar
+ Made multiple fixes and changes to Tiberon Prime

+++ wdfeerPlayer.cs
@@ -118,7 +118,7 @@ namespace wdfeerMod
         public int burstsMax = 2;
         public int burstCount = 1;
         int burstTimer = 0;
-        public Int64 longTimer = 0;
+        public int longTimer = 0;
         public ModItem burstItem;
         public override void PreUpdate()
         {
